[Feat] #23  피드 상세 화면 감정 표현 조회 및 추가 API 연동

+++ app/src/main/java/com/project/score/API/response/record/FeedDetailResponse.kt
@@ -1,7 +1,7 @@
 package com.project.score.API.response.record
 
 data class FeedDetailResponse(
-    val feedId: Long,
+    val feedId: Int,
     val uploaderNickname: String,
     val uploaderProfileImgUrl: String,
     val feedImg: String,
